Remove non-Proxy version
