+++ .travis.yml
@@ -1,5 +1,5 @@
 sudo: false
 language: node_js
 node_js:
+  - 'node'
   - '6'
-  - '4'
